+++ packages/generator-typescript/src/package-data-parser.ts
@@ -73,7 +73,7 @@ export class PackageDataParser {
 
     async start() {
         const filePath = resolve(__dirname, '../packages.xml')
-        console.log(`Parsing ${filePath}...`)
+        this.log.log(`Parsing ${filePath}...`)
         const fileContents = await readFile(filePath, 'utf8')
         const result = (await parseStringPromise(fileContents)) as ParsedPackageData
         this.parseSections(result.packages.section)